refactor(frontend): replace all discountPrice with price

+++ frontend/src/Admin/ProductsPage.jsx
@@ -1,5 +1,5 @@
 import { useEffect, useReducer, useState } from "react";
-import useFetch from "../utils/useFetch";
+import useFetch from "../utils/useFetch.js";
 import { FaRegEdit } from "react-icons/fa";
 import { MdOutlineDelete } from "react-icons/md";
 import { Link, useNavigate } from "react-router-dom";
@@ -7,11 +7,7 @@ import SearchField from "../components/SearchField";
 import getSearchResults from "../utils/getSearchResults";
 import { handleDelete } from "../utils/serverUtils";
 export default function ProductsPage() {
-  const {
-    data: products,
-    loading,
-    error,
-  } = useFetch("http://localhost:3000/products");
+  const { data: products, loading, error } = useFetch("http://localhost:3000/products");
 
   const navigate = useNavigate();
   const [displayProducts, setDisplayProducts] = useState(products);
@@ -23,16 +19,13 @@ export default function ProductsPage() {
     <div>
       <div className="flex items-end justify-between">
         <div>
-          <h1 className="text-2xl font-semibold text-slate-700 mb-3">
-            Products
-          </h1>
+          <h1 className="text-2xl font-semibold text-slate-700 mb-3">Products</h1>
           {loading && <p>Loading...</p>}
           {error && <p>Error: {error.message}</p>}
           {products && (
             <Link
               to="/admin/products/addproduct"
-              className="bg-slate-500 inline-block text-white px-2 py-1 rounded-md  gap-1 hover:bg-opacity-85 min-w-32 text-center"
-            >
+              className="bg-slate-500 inline-block text-white px-2 py-1 rounded-md  gap-1 hover:bg-opacity-85 min-w-32 text-center">
               Add Products
             </Link>
           )}
@@ -47,10 +40,7 @@ export default function ProductsPage() {
                 navigate(`/admin/products/${product.id}`);
               }}
             />
-            <Filter
-              setDisplayProducts={setDisplayProducts}
-              products={products}
-            />
+            <Filter setDisplayProducts={setDisplayProducts} products={products} />
           </div>
         )}
       </div>
@@ -59,8 +49,7 @@ export default function ProductsPage() {
           {displayProducts.map((product) => (
             <div
               key={product.id}
-              className="bg-slate-200 flex items-center border-b justify-between p-1 px-10 rounded-sm shadow-md"
-            >
+              className="bg-slate-200 flex items-center border-b justify-between p-1 px-10 rounded-sm shadow-md">
               <div>
                 <p className="text-xl font-semibold">{product.name}</p>
                 <img
@@ -68,29 +57,25 @@ export default function ProductsPage() {
                   alt="product image"
                   className="h-16 object-cover w-28 rounded-sm"
                 />
-                <p className="text-gray-600">Price: {product.discountPrice}</p>
+                <p className="text-gray-600">Price: {product.price}</p>
                 <p className="text-gray-500">Old Price: {product.oldPrice} </p>
                 <p className="text-gray-600">Category: {product.category} </p>
               </div>
               <div className="flex flex-col gap-1">
                 <Link
                   to={`/admin/products/${product.id}`}
-                  className="bg-slate-500 text-white px-2 py-1 rounded-md flex items-center gap-1 hover:bg-opacity-85"
-                >
+                  className="bg-slate-500 text-white px-2 py-1 rounded-md flex items-center gap-1 hover:bg-opacity-85">
                   Edit <FaRegEdit />
                 </Link>
                 <button
                   onClick={() => {
-                    window.confirm(
-                      "Are you sure you want to delete this item?"
-                    ) &&
+                    window.confirm("Are you sure you want to delete this item?") &&
                       handleDelete(product.id, "products") &&
                       setDisplayProducts((prev) =>
                         prev.filter((p) => p.id !== product.id)
                       );
                   }}
-                  className="bg-red-700 text-white px-2 py-1 rounded-md flex items-center gap-1 hover:bg-opacity-85"
-                >
+                  className="bg-red-700 text-white px-2 py-1 rounded-md flex items-center gap-1 hover:bg-opacity-85">
                   Delete <MdOutlineDelete />
                 </button>
               </div>
@@ -133,8 +118,7 @@ function Filter({ setDisplayProducts, products }) {
           name="category"
           id="category"
           value={filterOption}
-          onChange={(e) => setFilterOption(e.target.value)}
-        >
+          onChange={(e) => setFilterOption(e.target.value)}>
           <option value="">All</option>
           <option value="homedecor">Home Decor</option>
           <option value="sofas">Sofas</option>
@@ -146,8 +130,7 @@ function Filter({ setDisplayProducts, products }) {
       </div>
       <button
         className="bg-slate-500 text-white p-1 px-3 rounded-md w-full hover:bg-opacity-80"
-        onClick={handleFilter}
-      >
+        onClick={handleFilter}>
         Filter
       </button>
     </div>
